Update password auth to own package & rename passport-auth to auth-password
