+++ packages/auth-passport/lib/Facebook.js
