+++ packages/auth-passport/lib/GitHub.js
