+++ packages/auth-passport/lib/Passport.js
