+++ packages/auth-password/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@keystone-alpha/auth-password",
+  "description": "Provides a basic password authentication atrategy.",
+  "version": "0.0.0",
+  "author": "The KeystoneJS Development Team",
+  "license": "MIT",
+  "engines": {
+    "node": ">=8.4.0"
+  },
+  "dependencies": {
+    "@keystone-alpha/fields": "^10.0.0",
+    "@keystone-alpha/session": "^2.0.1"
+  }
+}
