+++ packages/keystone/index.js
@@ -1,9 +1,7 @@
 const Keystone = require('./lib/Keystone');
-const PasswordAuthStrategy = require('./lib/auth/Password');
 const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('./lib/adapters');
 module.exports = {
   Keystone,
-  PasswordAuthStrategy,
   BaseKeystoneAdapter,
   BaseListAdapter,
   BaseFieldAdapter,
